Dedup MSVC by splitting the atomic base class into a seperate header and creating back-end intrinsics in the MSVC header
