+++ libcxx/include/__config
@@ -1611,6 +1611,7 @@ _LIBCUDACXX_FUNC_VIS extern "C" void __sanitizer_annotate_contiguous_container(
 #  define _LIBCUDACXX_HAS_GCC_ATOMIC_IMP
 #elif defined(_LIBCUDACXX_COMPILER_MSVC)
 #  define _LIBCUDACXX_HAS_MSVC_ATOMIC_IMPL
+#  define _LIBCUDACXX_NO_RUNTIME_LOCK_FREE
 #endif
 
 // CUDA Atomics supersede host atomics in order to insert the host/device dispatch layer
